Big commit cuz version control got messed uo
