+++ Assets/Materials/New Material.mat
@@ -78,6 +78,6 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
-    - _Color: {r: 0.00002021609, g: 0, b: 0.066037714, a: 1}
+    - _Color: {r: 0.749199, g: 0.8679245, b: 0.7867011, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
   m_BuildTextureStacks: []
