+++ Assets/Scenes/cameraSettings.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 343b55b4ed9884c33a4fbc2dca80e57a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
