--- Assets/freeroam.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7dcaeb3ca1dc141438c429e5624a335c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
