--- Assets/startupscene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0ff6d522ebc894a07ba305d69d5f877b
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
